Make two methods static for testing, and stub out tests.
