+++ src/DynamoCoreWpf/ViewModels/Watch3D/HelixWatch3DViewModel.cs
@@ -1295,6 +1295,15 @@ namespace Dynamo.Wpf.ViewModels.Watch3D
             return mesh;
         }
 
+        internal void UpdateNearClipPlane()
+        {
+            var nearClipDistance = ComputeNearClipPlaneDistance(Camera.Position.ToVector3(), SceneItems,
+                NearPlaneDistanceFactor);
+
+            if (Camera.NearPlaneDistance == nearClipDistance) return;
+            Camera.NearPlaneDistance = nearClipDistance;
+        }
+
         /// <summary>
         /// This method attempts to maximize the near clip plane in order to 
         /// achiever higher z-buffer precision.
@@ -1303,30 +1312,25 @@ namespace Dynamo.Wpf.ViewModels.Watch3D
         /// of geometry bounding boxes in the scene . It sets the camera's near clip plane to 
         /// `near/2` and the far clip plane to `far`.
         /// </summary>
-        internal void UpdateNearClipPlane()
+        internal static double ComputeNearClipPlaneDistance(Vector3 cameraPosition, IEnumerable<Model3D> geometry, double nearPlaneDistanceFactor)
         {
             var closest = double.MaxValue;
             var farthest = double.MinValue;
 
-            var pos = Camera.Position.ToVector3();
-
             var centers =
-                SceneItems.Where(i => i is GeometryModel3D)
+                geometry.Where(i => i is GeometryModel3D)
                     .Cast<GeometryModel3D>()
                     .Select(g => (g.Bounds.Maximum + g.Bounds.Minimum)/2)
                     .ToList();
 
-            foreach (var d in centers.Select(c => Vector3.DistanceSquared(pos, c)))
+            foreach (var d in centers.Select(c => Vector3.DistanceSquared(cameraPosition, c)))
             {
                 closest = Math.Min(closest, d);
                 farthest = Math.Max(farthest, d);
             }
 
-            var nearClipDistance = Math.Min(Math.Sqrt(farthest)*NearPlaneDistanceFactor, Math.Sqrt(closest)/2);
-            if (Camera.NearPlaneDistance == nearClipDistance) return;
-
-            Camera.NearPlaneDistance = nearClipDistance;
             //Debug.WriteLine("Close={0}, Far={1}, Near clip={2}", Math.Sqrt(closest), Math.Sqrt(farthest), Camera.NearPlaneDistance);
+            return Math.Min(Math.Sqrt(farthest)* nearPlaneDistanceFactor, Math.Sqrt(closest)/2);
         }
 
         internal override void ExportToSTL(string path, string modelName)
@@ -1440,6 +1444,48 @@ namespace Dynamo.Wpf.ViewModels.Watch3D
             };
 
             return camData;
+        }
+    }
+
+    internal static class BoundingBoxExtensions
+    {
+        /// <summary>
+        /// Convert a <see cref="BoundingBox"/> to a <see cref="Rect3D"/>
+        /// </summary>
+        /// <param name="bounds">The <see cref="BoundingBox"/> to be converted.</param>
+        /// <returns>A <see cref="Rect3D"/> object.</returns>
+        public static Rect3D ToRect3D(this BoundingBox bounds)
+        {
+            var min = bounds.Minimum;
+            var max = bounds.Maximum;
+            var size = new Size3D((max.X - min.X), (max.Y - min.Y), (max.Z - min.Z));
+            return new Rect3D(min.ToPoint3D(), size);
+        }
+
+        /// <summary>
+        /// If a <see cref="GeometryModel3D"/> has more than one point, then
+        /// return its bounds, otherwise, return a unit-sized bounding
+        /// box surrounding the point.
+        /// </summary>
+        /// <param name="pointGeom">A <see cref="GeometryModel3D"/> object.</param>
+        /// <returns>A <see cref="BoundingBox"/> object encapsulating the geometry.</returns>
+        public static BoundingBox Bounds(this GeometryModel3D geom)
+        {
+            if (geom.Geometry.Positions.Count == 0)
+            {
+                return new BoundingBox();
+            }
+
+            if (geom.Geometry.Positions.Count > 1)
+            {
+                return BoundingBox.FromPoints(geom.Geometry.Positions.ToArray());
+            }
+
+            const float boundsSize = 5.0f;
+            var pos = geom.Geometry.Positions.First();
+            var min = pos + new Vector3(-boundsSize, -boundsSize, -boundsSize);
+            var max = pos + new Vector3(boundsSize, boundsSize, boundsSize);
+            return new BoundingBox(min, max);
         }
     }
 }
